[URP]Quality audit-Material editor help button (#4760)
* [URP]Adding a help button for showing the shader urls on the browser

* Changelog

* Fixes from Theo review

* Code review fixes

* Implemented suggestions for docs from Kerry

* Merge fixes

+++ com.unity.render-pipelines.universal/Editor/ShaderGUI/Shaders/LitShader.cs
@@ -13,7 +13,7 @@ namespace UnityEditor.Rendering.Universal.ShaderGUI
 
         public override void FillAdditionalFoldouts(MaterialHeaderScopeList materialScopesList)
         {
-            materialScopesList.RegisterHeaderScope(LitDetailGUI.Styles.detailInputs, (uint)Expandable.Details, _ => LitDetailGUI.DoDetailArea(litDetailProperties, materialEditor));
+            materialScopesList.RegisterHeaderScope(LitDetailGUI.Styles.detailInputs, Expandable.Details, _ => LitDetailGUI.DoDetailArea(litDetailProperties, materialEditor));
         }
 
         // collect properties from the material properties
